Disable Generate button when background minimization isn't set correctly
